#22 - Make waiting for detachment to succeed optional via ENV var `ASG_WAIT_FOR_DETACHMENT`
This can be useful for cases when detachment is extremely slow; e.g draining connections off an NLB and target group. Even though the instance has been terminated and is unhealthy, this can take 1-2 minutes even with reduced `deregistration_delay`. This allows people to opt-in to choose not to wait for the detachment to complete.